Last group Project
 Java program to solve the following synchronization & deadlock problem.

Two mountain resorts are connected by a road that becomes single lane inside a tunnel that goes through the
mountain. One resort is at the left end of the tunnel and the other at the right end of the tunnel. Cars moving
from left to right are called “Right-bound” cars and are indicated with odd numbers starting from 1 (1, 3, 5, …).
“Left-bound” cars move from right to left and are indicated with even numbers starting from 0 (0, 2, 4, ..). The
tunnel can become deadlocked if a left-bound and a right-bound car enter into the tunnel at the same time.